improvements to line graph
